Move asnative implementations
